Fixed warnings from clippy
